ouroboros-network: introduce Util.gshowsPrecForgetFields
This function derives Show instances corresponding to the same types without
record syntax.

This commit also changes a few instances, those I've seen often when debugging
the consensus tests. It may be useful to change other Show instances as well.

The implementation of Util.gshowsPrecForgetFields could be simplified somewhat
at the cost of some redundancy: the Coercible constraint could be replaced with
a class internal to this module that builds up the necessary Coercion.