Protect against unintended file deletion
Addresses #445. The ChainDB deletes invalid files on startup. When
accidentally pointing it to the database directory of another net (testnet vs.
mainnet), it would wipe out that directory, throwing away all blocks. To
protect the user against such accidents, we store the `ProtocolMagicId` in a
file in the database folder and check that upon startup.

See the docstring of `checkProtocolMagicId` for more details.